refactor(mf2): Rearrange exports, adding messageformat/functions & messageformat/functions/utils
