+++ packages/mf2-messageformat/package.json
@@ -25,6 +25,18 @@
   "files": [
     "lib/"
   ],
+  "exports": {
+    ".": {
+      "default": "./lib/index.js"
+    },
+    "./functions": {
+      "default": "./lib/functions/index.js"
+    },
+    "./functions/utils": {
+      "default": "./lib/functions/utils.js"
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
     "build": "tsc --project tsconfig.build.json",
     "extract-api": "api-extractor run --local --verbose"
