+++ packages/mf2-messageformat/src/index.ts
@@ -3,21 +3,30 @@ export type * from './formatted-parts.js';
 export type * from './functions/index.js';
 
 export { asDataModel, parseMessage, type CST } from './cst-parser/index.js';
-export * from './data-model/type-guards.js';
+export {
+  isCatchallKey,
+  isExpression,
+  isFunctionAnnotation,
+  isLiteral,
+  isMarkup,
+  isMessage,
+  isPatternMessage,
+  isSelectMessage,
+  isUnsupportedAnnotation,
+  isVariableRef
+} from './data-model/type-guards.js';
 export { validate } from './data-model/validate.js';
 export { visit } from './data-model/visit.js';
 export {
   MessageDataModelError,
   MessageError,
   MessageResolutionError,
+  MessageSelectionError,
   MessageSyntaxError
 } from './errors.js';
-export { defaultFunctions, utilFunctions } from './functions/index.js';
 export {
-  asBoolean,
-  asPositiveInteger,
-  asString,
-  mergeLocales
-} from './functions/utils.js';
-export { MessageFormat, MessageFormatOptions } from './messageformat.js';
+  MessageFormat,
+  MessageFormatOptions,
+  MessageFunctions
+} from './messageformat.js';
 export { stringifyMessage } from './stringifier/message.js';
