+++ tsconfig.json
@@ -15,7 +15,14 @@
       "@messageformat/react": ["packages/react/src/index"],
       "@messageformat/runtime": ["packages/runtime/src/runtime"],
       "@messageformat/runtime/lib/*": ["packages/runtime/src/*"],
-      "messageformat": ["packages/mf2-messageformat/src/index.ts"]
+      "@messageformat/test-utils": ["packages/test-utils/src/index.ts"],
+      "messageformat": ["packages/mf2-messageformat/src/index.ts"],
+      "messageformat/functions": [
+        "packages/mf2-messageformat/src/functions/index.ts"
+      ],
+      "messageformat/functions/utils": [
+        "packages/mf2-messageformat/src/functions/utils.ts"
+      ]
     }
   }
 }
